requirements.txt
browser-use
playwright
python-dotenv


